The tab test page was edited and some bugs were fixed and security improvements were made.

+++ private/views/test-tab-view.inc.php
@@ -104,13 +104,17 @@
 
                 <p class="card-text float-end">
 
-                    <a href="<?= ROOT ?>/single_test/editquestion/<?= $row->test_id ?>/<?= $question->id ?><?= $type ?>">
-                        <button class="btn btn-info text-white pe-1"><i class="fa fa-edit"></i></button>
-                    </a>
+                    <?php if ($row->editable): ?>
 
-                    <a href="<?= ROOT ?>/single_test/deletequestion/<?= $row->test_id ?>/<?= $question->id ?><?= $type ?>">
-                        <button class="btn btn-danger text-white pe-1"><i class="fa fa-trash-alt"></i></button>
-                    </a>
+                        <a href="<?= ROOT ?>/single_test/editquestion/<?= $row->test_id ?>/<?= $question->id ?><?= $type ?>">
+                            <button class="btn btn-info text-white pe-1"><i class="fa fa-edit"></i></button>
+                        </a>
+
+                        <a href="<?= ROOT ?>/single_test/deletequestion/<?= $row->test_id ?>/<?= $question->id ?><?= $type ?>">
+                            <button class="btn btn-danger text-white pe-1"><i class="fa fa-trash-alt"></i></button>
+                        </a>
+
+                    <?php endif; ?>
                 </p>
             </div>
         </div>
